Create vn-most.json
I have provided as much information as possible. @kd-ods  I also shared an example certificate with the registration number via email. Please let me know if it will be possible to use the VN-MOST as an identifier and if additional information is needed. First time submitting the JSON file so I am sure there will be errors for me to fix :)