Updated vn-most.json
@kd-ods thank you so much for all the helpful feedback. I have now done all the relevant edits and will commit directly to the same branch.

On the point for "subnationalCoverage", I still need to make an edit. You are correct that it is not only government agencies. I couldn't identify if there was a codelist for that or shall I just type in "research organisations". Apologies if that was listed somewhere and I missed it! Thanks so much again.

  "subnationalCoverage": [],
  "structure": [
    "government_agency"